+++ packages/ember-auto-import/babel-plugin/index.js
@@ -1,20 +1,2 @@
-const template  = require('babel-template');
-const syntax = require('babel-plugin-syntax-dynamic-import');
-const buildImport = template(`emberAutoImportDynamic(SOURCE)`);
-
-module.exports = () => ({
-  inherits: syntax,
-  visitor: {
-    Import(path) {
-      const newImport = buildImport({
-        SOURCE: path.parentPath.node.arguments,
-      });
-      path.parentPath.replaceWith(newImport);
-    },
-  },
-});
-
-module.exports.baseDir = function() {
-  return __dirname;
-};
-
+// eslint-disable-next-line
+module.exports = require('../js/babel-plugin');