+++ packages/ember-auto-import/ts/bundle-config.ts
@@ -4,23 +4,33 @@
 */
 
 import { dirname } from 'path';
+import { AppInstance } from './ember-cli-models';
 const testsPattern = new RegExp(`^/?[^/]+/(tests|test-support)/`);
 
+import type { TreeType } from './analyzer';
+
+function exhausted(label: string, value: never): never {
+  throw new Error(`Unknown ${label} specified: ${value}`);
+}
+
+export type BundleName = 'app' | 'tests';
+export type BundleType = 'js' | 'css';
+
 export default class BundleConfig {
-  constructor(private emberApp: any) {}
+  constructor(private emberApp: AppInstance) {}
 
   // This list of valid bundles, in priority order. The first one in the list that
   // needs a given import will end up with that import.
-  get names() : ReadonlyArray<string> {
+  get names() : ReadonlyArray<BundleName> {
     return Object.freeze(['app', 'tests']);
   }
 
-  get types(): ReadonlyArray<string> {
+  get types(): ReadonlyArray<BundleType> {
     return Object.freeze(['js', 'css']);
   }
 
   // Which final JS file the given bundle's dependencies should go into.
-  bundleEntrypoint(name: string, type: string): string | undefined {
+  bundleEntrypoint(name: BundleName, type: BundleType): string | undefined {
     switch (name) {
       case 'tests':
         switch (type) {
@@ -28,6 +38,8 @@ export default class BundleConfig {
             return 'assets/test-support.js';
           case 'css':
             return 'assets/test-support.css';
+          default:
+            exhausted('test bundle type', type);
         }
       case 'app':
         switch (type) {
@@ -35,13 +47,35 @@ export default class BundleConfig {
             return this.emberApp.options.outputPaths.vendor.js.replace(/^\//, '');
           case 'css':
             return this.emberApp.options.outputPaths.vendor.css.replace(/^\//, '');
+          default:
+            exhausted('app bundle type', type);
         }
+      default:
+        exhausted('bundle name', name);
+    }
+  }
+
+  bundleForTreeType(treeType: TreeType): BundleName {
+    switch (treeType) {
+      case 'app':
+      case 'addon':
+      case 'addon-templates':
+      case 'styles':
+      case 'templates':
+        return 'app';
+
+      case 'addon-test-support':
+      case 'test':
+        return 'tests';
+
+      default:
+        exhausted('bundle name', treeType);
     }
   }
 
   // For any relative path to a module in our application, return which bundle its
   // imports go into.
-  bundleForPath(path: string): string {
+  bundleForPath(path: string): BundleName {
     if (testsPattern.test(path)) {
       return 'tests';
     } else {