+++ packages/ember-auto-import/ts/package.ts
@@ -1,10 +1,11 @@
-import resolve from 'resolve';
+import resolvePackagePath from 'resolve-package-path';
 import { join } from 'path';
 import { readFileSync } from 'fs';
 import { Memoize } from 'typescript-memoize';
 import { Configuration } from 'webpack';
+import { AddonInstance, isDeepAddonInstance, Project } from './ember-cli-models';
 
-const cache: WeakMap<any, Package> = new WeakMap();
+const cache: WeakMap<AddonInstance, Package> = new WeakMap();
 let pkgGeneration = 0;
 
 export function reloadDevPackages() {
@@ -20,45 +21,73 @@ export interface Options {
   skipBabel?: { package: string, semverRange?: string }[];
 }
 
+interface DepResolution {
+  type: 'package';
+  path: string;
+  packageName: string;
+  packagePath: string;
+  local: string;
+}
+
+interface LocalResolution {
+  type: 'local';
+  local: string;
+}
+
+interface URLResolution {
+  type: 'url';
+  url: string;
+}
+
+interface ImpreciseResolution {
+  type: 'imprecise';
+}
+
+type Resolution = DepResolution | LocalResolution | URLResolution | ImpreciseResolution;
+
 export default class Package {
   public name: string;
   public root: string;
   public isAddon: boolean;
   private _options: any;
-  private _parent: any;
+  private _parent: Project | AddonInstance;
   private _hasBabelDetails = false;
   private _babelMajorVersion?: number;
   private _babelOptions: any;
   private _emberCLIBabelExtensions?: string[];
   private autoImportOptions: Options | undefined;
-  private isAddonCache = new Map<string, boolean>();
   private isDeveloping: boolean;
   private pkgGeneration: number;
   private pkgCache: any;
 
-  static lookup(appOrAddon: any): Package {
-    if (!cache.has(appOrAddon)) {
-      cache.set(appOrAddon, new this(appOrAddon));
+  static lookupParentOf(child: AddonInstance): Package {
+    if (!cache.has(child)) {
+      cache.set(child, new this(child));
     }
-    return cache.get(appOrAddon)!;
+    return cache.get(child)!;
   }
 
-  constructor(appOrAddon: any) {
-    this.name = appOrAddon.parent.pkg.name;
-    this.root = appOrAddon.parent.root;
-    this.isAddon = appOrAddon.parent !== appOrAddon.project;
-    this.isDeveloping = !this.isAddon || this.root === appOrAddon.project.root;
+  constructor(child: AddonInstance) {
+    this.name = child.parent.pkg.name;
+    this.root = child.parent.root;
 
-    // This is the per-package options from ember-cli
-    this._options = this.isAddon
-      ? appOrAddon.parent.options
-      : appOrAddon.app.options;
-    this._parent = appOrAddon.parent;
+    if (isDeepAddonInstance(child)) {
+      this.isAddon = true;
+      this.isDeveloping = this.root === child.project.root;
+      // This is the per-package options from ember-cli
+      this._options = child.parent.options;
+    } else {
+      this.isAddon = false;
+      this.isDeveloping = true;
+      this._options = child.app.options;
+    }
+
+    this._parent = child.parent;
 
     // Stash our own config options
     this.autoImportOptions = this._options.autoImport;
 
-    this.pkgCache = appOrAddon.parent.pkg;
+    this.pkgCache = child.parent.pkg;
     this.pkgGeneration = pkgGeneration;
   }
 
@@ -86,17 +115,17 @@ export default class Package {
   get isFastBootEnabled() {
     return process.env.FASTBOOT_DISABLED !== 'true'
     && !!this._parent.addons.find(
-      (addon: any) => addon.name === 'ember-cli-fastboot'
+      addon => addon.name === 'ember-cli-fastboot'
     );
   }
 
-  private buildBabelOptions(instance: any, options: any) {
+  private buildBabelOptions(instance: Project | AddonInstance, options: any) {
     // Generate the same babel options that the package (meaning app or addon)
     // is using. We will use these so we can configure our parser to
     // match.
     let babelAddon = instance.addons.find(
-      (addon: any) => addon.name === 'ember-cli-babel'
-    );
+      addon => addon.name === 'ember-cli-babel'
+    ) as any;
     let babelOptions = babelAddon.buildBabelOptions(options);
     let extensions = babelOptions.filterExtensions || ['js'];
 
@@ -134,7 +163,7 @@ export default class Package {
     return `${this.name}/${this.isAddon ? 'addon' : 'app'}`;
   }
 
-  hasDependency(name: string): boolean {
+  private hasDependency(name: string): boolean {
     let pkg = this.pkg;
     return (
       (pkg.dependencies && Boolean(pkg.dependencies[name])) ||
@@ -151,18 +180,76 @@ export default class Package {
     );
   }
 
-  isEmberAddonDependency(name: string): boolean {
-    if (!this.isAddonCache.has(name)) {
-      let packageJSON = require(resolve.sync(`${name}/package.json`, {
-        basedir: this.root
-      }));
-      let isV1Addon = packageJSON.keywords?.includes('ember-addon') && packageJSON['ember-addon']?.version !== 2;
-      this.isAddonCache.set(name, isV1Addon);
+  static categorize(importedPath: string, partial = false) {
+    if (/^(\w+:)?\/\//.test(importedPath)) {
+      return 'url';
     }
-    return this.isAddonCache.get(name) || false;
+
+    if (importedPath[0] === '.' || importedPath[0] === '/') {
+      return 'local';
+    }
+
+    if (partial && !isPrecise(importedPath)) {
+      return 'imprecise';
+    }
+    return 'dep';
   }
 
-  assertAllowedDependency(name: string) {
+  resolve(importedPath: string): DepResolution | LocalResolution | URLResolution;
+  resolve(importedPath: string, partial: true): DepResolution | LocalResolution | URLResolution | ImpreciseResolution;
+  resolve(importedPath: string, partial=false): Resolution | undefined {
+    switch (Package.categorize(importedPath, partial)) {
+      case 'url':
+        // unambiguous URLs with a scheme are allowed but ignored by us
+        if (/^(\w+:)?\/\//.test(importedPath)) {
+          return { type: "url", url: importedPath };
+        }
+      case 'local':
+        return {
+          type: "local",
+          local: importedPath
+        };
+      case 'imprecise':
+        if (partial) {
+          return {
+            type: "imprecise",
+          };
+        }
+        break;
+    }
+
+    let path = this.aliasFor(importedPath);
+    let [first, ...rest] = path.split('/');
+    let packageName;
+    if (first[0] === '@') {
+      packageName = `${first}/${rest.shift()}`;
+    } else {
+      packageName = first;
+    }
+
+    if (this.excludesDependency(packageName)) {
+      // This package has been explicitly excluded.
+      return;
+    }
+
+    if (!this.hasDependency(packageName)) {
+      return;
+    }
+
+    let packagePath = resolvePackagePath(packageName, this.root);
+    if (packagePath === null) {
+      throw new Error(`${ this.name } tried to import "${packageName}" but the package was not resolvable from ${this.root}`);
+    }
+
+    if (isEmberAddonDependency(packagePath)) {
+      // ember addon are not auto imported
+      return;
+    }
+    this.assertAllowedDependency(packageName);
+    return { type: 'package', path, packageName, local: rest.join('/'), packagePath };
+  }
+
+  private assertAllowedDependency(name: string) {
     if (this.isAddon && !this.hasNonDevDependency(name)) {
       throw new Error(
         `${
@@ -172,7 +259,7 @@ export default class Package {
     }
   }
 
-  excludesDependency(name: string): boolean {
+  private excludesDependency(name: string): boolean {
     return Boolean(
       this.autoImportOptions &&
       this.autoImportOptions.exclude &&
@@ -188,7 +275,7 @@ export default class Package {
     return this.autoImportOptions && this.autoImportOptions.skipBabel;
   }
 
-  aliasFor(name: string): string {
+  private aliasFor(name: string): string {
     return (
       (this.autoImportOptions &&
         this.autoImportOptions.alias &&
@@ -221,4 +308,30 @@ export default class Package {
       !this.isAddon && this.autoImportOptions && this.autoImportOptions.forbidEval
     );
   }
+}
+
+const isAddonCache = new Map<string, boolean>();
+function isEmberAddonDependency(pathToPackageJSON: string): boolean {
+  let cached = isAddonCache.get(pathToPackageJSON);
+  if (cached === undefined) {
+    let packageJSON = require(pathToPackageJSON);
+    let answer = packageJSON.keywords?.includes('ember-addon') && packageJSON['ember-addon']?.version !== 2;
+    isAddonCache.set(pathToPackageJSON, answer);
+    return answer;
+  } else {
+    return cached;
+  }
+}
+
+function count(str: string, letter: string): number {
+  return [...str].reduce((a,b) => a + (b === letter ? 1 : 0), 0);
+}
+
+function isPrecise(leadingQuasi: string): boolean {
+  if (leadingQuasi.startsWith('.') || leadingQuasi.startsWith('/')) {
+    return true;
+  }
+  let slashes = count(leadingQuasi, '/');
+  let minSlashes = leadingQuasi.startsWith('@') ? 2 : 1;
+  return slashes >= minSlashes;
 }