+++ packages/ember-auto-import/ts/tests/analyzer-test.ts
@@ -1,10 +1,11 @@
 import QUnit from 'qunit';
+import 'qunit-assertions-extra';
 import broccoli, { Builder } from 'broccoli';
 import { UnwatchedDir } from 'broccoli-source';
 import quickTemp from 'quick-temp';
 import { ensureDirSync, readFileSync, outputFileSync, removeSync, existsSync } from 'fs-extra';
 import { join } from 'path';
-import Package from '../package';
+import type Package from "../package";
 import Analyzer from '../analyzer';
 
 const { module: Qmodule, test } = QUnit;
@@ -14,7 +15,7 @@ Qmodule('analyzer', function(hooks) {
   let builder: Builder;
   let upstream: string;
   let analyzer: Analyzer;
-  let pack: Partial<Package>;
+  let pack: Package;
   let babelOptionsWasAccessed = false;
 
   hooks.beforeEach(function(this: any) {
@@ -23,12 +24,14 @@ Qmodule('analyzer', function(hooks) {
     pack = {
       get babelOptions() {
         babelOptionsWasAccessed = true;
-        return {};
+        return {
+          plugins: [require.resolve('../../babel-plugin')]
+        };
       },
       babelMajorVersion: 6,
-      fileExtensions: ['js']
-    };
-    analyzer = new Analyzer(new UnwatchedDir(upstream), pack as Package);
+      fileExtensions: ["js"],
+    } as Package;
+    analyzer = new Analyzer(new UnwatchedDir(upstream), pack);
     builder = new broccoli.Builder(analyzer);
   });
 
@@ -96,7 +99,8 @@ Qmodule('analyzer', function(hooks) {
       isDynamic: false,
       specifier: 'some-package',
       path: 'sample.js',
-      package: pack
+      package: pack,
+      treeType: undefined,
     }]);
   });
 
@@ -113,7 +117,8 @@ Qmodule('analyzer', function(hooks) {
       isDynamic: false,
       specifier: 'some-package',
       path: 'sample.js',
-      package: pack
+      package: pack,
+      treeType: undefined,
     }]);
   });
 
@@ -130,12 +135,14 @@ Qmodule('analyzer', function(hooks) {
       isDynamic: false,
       specifier: 'some-package',
       path: 'sample.js',
-      package: pack
+      package: pack,
+      treeType: undefined,
     },{
       isDynamic: false,
       specifier: 'other-package',
       path: 'sample.js',
-      package: pack
+      package: pack,
+      treeType: undefined,
     }]);
   });
 
@@ -162,4 +169,102 @@ Qmodule('analyzer', function(hooks) {
     assert.deepEqual(analyzer.imports, []);
   });
 
+  type LiteralExample = [string, string];
+  type TemplateExample = [string, string[], string[]];
+  function isLiteralExample(exp: LiteralExample | TemplateExample): exp is LiteralExample {
+    return exp.length === 2;
+  }
+
+  let legalDyamicExamples: (LiteralExample | TemplateExample)[] = [
+    ["import('alpha');", "alpha"],
+    ["import('@beta/thing');", "@beta/thing"],
+    ["import(`gamma`);", "gamma"],
+    ["import(`@delta/thing`);", "@delta/thing"],
+    ["import('epsilon/mod');", "epsilon/mod"],
+    ["import('@zeta/thing/mod');", "@zeta/thing/mod"],
+    ["import(`eta/mod`);", "eta/mod"],
+    ["import(`@theta/thing/mod`);", "@theta/thing/mod"],
+    ["import('http://example.com');", "http://example.com"],
+    ["import('https://example.com');", "https://example.com"],
+    ["import('//example.com');", "//example.com"],
+    ["import(`http://example.com`);", "http://example.com"],
+    ["import(`https://example.com`);", "https://example.com"],
+    ["import(`//example.com`);", "//example.com"],
+    ["import(`http://example.com`);", "http://example.com"],
+    ["import(`https://example.com`);", "https://example.com"],
+    ["import(`//example.com`);", "//example.com"],
+    ["import(`http://${domain}`);", ["http://", ""], ["domain"]],
+    [
+      "import(`https://example.com/${path}`);",
+      ["https://example.com/", ""],
+      ["path"],
+    ],
+    ["import(`//${domain}`);", ["//", ""], ["domain"]],
+    ["import(`alpha/${foo}`);", ["alpha/", ""], ["foo"]],
+    ["import(`@beta/thing/${foo}`);", ["@beta/thing/", ""], ["foo"]],
+    ["import(`alpha/${foo}/component`);", ["alpha/", "/component"], ["foo"]],
+    [
+      "import(`@beta/thing/${foo}/component`);",
+      ["@beta/thing/", "/component"],
+      ["foo"],
+    ],
+    [
+      "import(`alpha/${foo}/component/${bar}`);",
+      ["alpha/", "/component/", ""],
+      ["foo", "bar"],
+    ],
+    [
+      "import(`@beta/thing/${foo}/component/${bar}`);",
+      ["@beta/thing/", "/component/", ""],
+      ["foo", "bar"],
+    ],
+  ];
+
+  for (let example of legalDyamicExamples) {
+    let [src] = example;
+    test(`dynamic import example: ${src}`, async function (assert) {
+      outputFileSync(join(upstream, "sample.js"), src);
+      await builder.build();
+      if (isLiteralExample(example)) {
+        assert.deepEqual(analyzer.imports, [
+          {
+            isDynamic: true,
+            specifier: example[1],
+            path: "sample.js",
+            package: pack,
+            treeType: undefined,
+          },
+        ]);
+      } else {
+        assert.deepEqual(analyzer.imports, [
+          {
+            cookedQuasis: example[1],
+            expressionNameHints: example[2],
+            path: "sample.js",
+            package: pack,
+            treeType: undefined,
+          },
+        ]);
+      }
+    });
+  }
+
+  test("disallowed patttern: unsupported syntax", async function (assert) {
+    assert.expect(1);
+    let src = `
+    function x() {
+      import((function(){ return 'hi' })());
+    }
+    `;
+    outputFileSync(join(upstream, "sample.js"), src);
+    try {
+      await builder.build();
+      throw new Error(`expected not to get here, build was supposed to fail`);
+    } catch (err) {
+      assert.contains(
+        err.message,
+        "import() is only allowed to contain string literals or template string literals"
+      );
+    }
+  });
 });