+++ packages/sample-direct/tests/acceptance/dynamic-import-test.js
@@ -10,4 +10,18 @@ module('Acceptance | dynamic import', function(hooks) {
     assert.equal(currentURL(), '/dynamic-import');
     assert.equal(document.querySelector('[data-test="dynamic-import-result"]').textContent.trim(), 'ember-auto-import-a-dependency');
   });
+
+  test('template dynamic-import', async function(assert) {
+    await visit('/flavor/vanilla');
+    assert.equal(currentURL(), '/flavor/vanilla');
+    assert.equal(document.querySelector('[data-test="dynamic-import-result"]').textContent.trim(), 'vanilla');
+  });
+
+  test('browser can use native import', async function(assert) {
+    await visit('/native-import');
+    assert.equal(currentURL(), '/native-import');
+    let expected = typeof FastBoot === 'undefined' ? 'browser' : 'server';
+    assert.equal(document.querySelector('[data-test="dynamic-import-result"]').textContent.trim(), expected);
+  });
+
 });