+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 1.9.0
+ - ENHANCEMENT: use new API from ember-cli to reliably detect which trees are test-support only, even when addons override the default naming scheme by @rwjblue
+ - ENHANCEMENT: switch to resolve-package-path for better shared caching with the rest of ember-cli by @rwjblue
+
+### 1.8.0
+ - ENHANCEMENT: improved leader election protocol between copies of ember-auto-import that ensures the newest one is always in charge.
+ - HOUSEKEEPING: upgrades to typescript and some other deps to get better upstream types
+
 ### 1.7.0
  - DOCS: improvement to CONTRIBUTING.md by kiwiupover
  - BUGFIX: fix merging of webpack configs by @bendemboski